# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel V12 Kanban Task Management API - FSJ28",
        "description": "API documentation for the Kanban task management system. Laravel 12.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1 Base URL"
        }
    ],
    "paths": {
...
        "/user/tasks": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get tasks for authenticated user",
                "description": "Retrieves all tasks assigned to the currently authenticated user. Requires authentication.",
                "operationId": "66a4f5ac4b7490eb33ccbf155ea920dd",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Missing or invalid token)"
                    },
                    "404": {
                        "description": "No tasks found for the authenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No tasks found for the authenticated user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "title": "Task",
                "description": "Task model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Implement Task API"
                    },
                    "description": {
                        "type": "string",
                        "example": "Develop CRUD operations for the Task resource."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "en proceso",
                            "completada"
                        ],
                        "example": "en proceso"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "baja",
                            "media",
                            "alta"
                        ],
                        "example": "alta"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-30"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "user": {
                        "description": "Associated user name (only in index/list view)",
                        "type": "string",
                        "readOnly": "true",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the Bearer token (JWT) returned at login.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}